move global alert into common koa-middleware
